refactor(canvas): colocate chat message status
