+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatMessages/index.tsx
@@ -7,19 +7,19 @@ import {
   type KeyboardEvent,
   type MouseEvent,
 } from 'react';
-import type { AgentChatMessage, CanvasNode } from '../../types';
-import { BotAvatarIcon } from '../icons';
-import { ChatMessage } from './ChatMessage';
-import { ChatActivityStatus } from './ChatActivityStatus';
-import { ChatThreadSkeleton } from './ChatThreadSkeleton';
-import type { PendingClarification, ToolCallStatus } from './types';
-import { buildAnchorElementId } from './utils/anchors';
-import { useI18n } from '../../i18n';
-import { isVSCodeLink } from './utils/externalLinks';
-import { localPathFromHref } from './utils/localFileLinks';
-import { ChatClarificationCard } from './ChatClarificationCard';
-import { useChatMessagesStatus } from './hooks/useChatMessagesStatus';
-import { tabRefFromMentionElement } from './utils/tabMentions';
+import type { AgentChatMessage, CanvasNode } from '../../../types';
+import { BotAvatarIcon } from '../../icons';
+import { ChatMessage } from '../ChatMessage';
+import { ChatActivityStatus } from '../ChatActivityStatus';
+import { ChatThreadSkeleton } from '../ChatThreadSkeleton';
+import type { PendingClarification, ToolCallStatus } from '../types';
+import { buildAnchorElementId } from '../utils/anchors';
+import { useI18n } from '../../../i18n';
+import { isVSCodeLink } from '../utils/externalLinks';
+import { localPathFromHref } from '../utils/localFileLinks';
+import { ChatClarificationCard } from '../ChatClarificationCard';
+import { useChatMessagesStatus } from './useChatMessagesStatus';
+import { tabRefFromMentionElement } from '../utils/tabMentions';
 
 /** How close (px) to the bottom still counts as "reading the tail" — within
  *  this band the view keeps following the stream; beyond it the user has
